pacthed save oof products. turning on refreshing the interface. fixed error when computing summary values on products update
P.S.
ok, this stack is i never worked with, so i believe lots of stuff can be made in nicer way, and i look like junior.
probablly Form object was ment to be cerate form directly and his fields as well. but, in this case there will be two different approaches how we cerate content in the page, and i don't like it. so i modified how the Form object is working - to use one language for templating. Thus i had to dig deep into libraries.
For the product table, of course it will be better to have each rows as it's own object, i had a prototype but was never able to render the chunk of the row into the table, thus, the row is part pf the products table object.

anyway as i see, i f@**kd it up, i'm looking as a lamer, but i did my best.